requirements.txt
discord.py==1.4.1
python-dotenv==0.15.0